+++ modio/client.py
@@ -6,11 +6,15 @@ import asyncio
 import datetime
 import logging
 import time
+from typing import Optional
 import aiohttp
 import requests
 
+from modio.utils import async_ratelimit_retry, ratelimit_retry
+
 from .errors import modioException
 from .entities import Event, Message, ModFile, Rating, User
+from .enums import TargetPlatform, TargetPortal
 from .objects import Pagination, Returned, Filter
 from .game import Game
 from .mod import Mod
@@ -19,12 +23,14 @@ from .mod import Mod
 class Connection:
     """Class handling under the hood requests and ratelimits."""
 
-    def __init__(self, api_key, access_token, lang, version, test):
+    def __init__(self, api_key, access_token, lang, version, test, platform, portal):
         self.test = test
         self.version = version
         self.access_token = access_token
         self.api_key = api_key
         self.lang = lang
+        self.platform = platform
+        self.portal = portal
 
         self.rate_limit = None
         self.rate_remain = None
@@ -116,9 +122,15 @@ class Connection:
                 "Content-Type": "application/x-www-form-urlencoded",
             }
 
+        if self.platform is not None:
+            headers["X-Modio-Platform"] = self.platform.value
+
+        if self.portal is not None:
+            headers["X-Modio-Portal"] = self.portal.value
+
         return headers
 
-    def _post(self, resp):
+    def _post_process(self, resp):
         try:
             resp_json = resp.json()
         except requests.JSONDecodeError:
@@ -134,6 +146,7 @@ class Connection:
 
         return data
 
+    @ratelimit_retry(2)
     def get_request(self, url, *, h_type=0, **fields):
         filters = fields.pop("filters", None)
         filters = (filters or Filter()).get_dict()
@@ -145,21 +158,24 @@ class Connection:
             h_type = 2
 
         resp = self.session.get(self._base_path + url, headers=self._define_headers(h_type), params=extra)
-        return self._post(resp)
+        return self._post_process(resp)
 
+    @ratelimit_retry(2)
     def post_request(self, url, *, h_type=0, **fields):
         resp = self.session.post(self._base_path + url, headers=self._define_headers(h_type), **fields)
-        return self._post(resp)
+        return self._post_process(resp)
 
+    @ratelimit_retry(2)
     def put_request(self, url, *, h_type=0, **fields):
         resp = self.session.put(self._base_path + url, headers=self._define_headers(h_type), **fields)
-        return self._post(resp)
+        return self._post_process(resp)
 
+    @ratelimit_retry(2)
     def delete_request(self, url, *, h_type=0, **fields):
         resp = self.session.delete(self._base_path + url, headers=self._define_headers(h_type), **fields)
-        return self._post(resp)
+        return self._post_process(resp)
 
-    async def _async_post(self, resp):
+    async def _async_post_process(self, resp):
         try:
             resp_json = await resp.json()
         except aiohttp.ContentTypeError:
@@ -175,6 +191,7 @@ class Connection:
 
         return data
 
+    @async_ratelimit_retry(2)
     async def async_get_request(self, url, *, h_type=0, **fields):
         filters = fields.pop("filters", None)
         filters = (filters or Filter()).get_dict()
@@ -188,8 +205,9 @@ class Connection:
         async with self.async_session.get(
             self._base_path + url, headers=self._define_headers(h_type), params=extra
         ) as resp:
-            return await self._async_post(resp)
+            return await self._async_post_process(resp)
 
+    @async_ratelimit_retry(2)
     async def async_post_request(self, url, *, h_type=0, **fields):
         files = fields.pop("files", {})
         data = fields.pop("data", {})
@@ -213,19 +231,21 @@ class Connection:
         async with self.async_session.post(
             self._base_path + url, headers=self._define_headers(h_type), data=form
         ) as resp:
-            return await self._async_post(resp)
+            return await self._async_post_process(resp)
 
+    @async_ratelimit_retry(2)
     async def async_put_request(self, url, *, h_type=0, **fields):
         async with self.async_session.put(
             self._base_path + url, headers=self._define_headers(h_type), **fields
         ) as resp:
-            return await self._async_post(resp)
+            return await self._async_post_process(resp)
 
+    @async_ratelimit_retry(2)
     async def async_delete_request(self, url, *, h_type=0, **fields):
         async with self.async_session.delete(
             self._base_path + url, headers=self._define_headers(h_type), **fields
         ) as resp:
-            return await self._async_post(resp)
+            return await self._async_post_process(resp)
 
 
 class Client:
@@ -252,25 +272,40 @@ class Client:
     version : Optional[str]
         An optional keyword argument to allow you to pick a specific version of the API to query,
         usually you shouldn't need to change this. Default is the latest supported version.
+    platform : Optiona[TargetPlatform]
+        The platform to target with requests.
+    portal : Optional[TargetPortal]
+        The portal to target with requests.
 
     Attributes
     -----------
-    rate_limit : int
-        Number of requests that can be made using the supplied API Key/access token.
-    rate_remain : int
-        Number of requests remaining. Once this number hits 0 the requests will become
-        rejected and the library will sleep until the limit resets then raise 429 TooManyRequests.
     retry_after : int
         Number of seconds until the rate limits are reset for this API Key/access token.
-        Is 0 until the rate_remain is 0 and becomes 0 again once the rate limit is reset.
+        Is 0 until the API returns a 429.
     """
 
-    def __init__(self, *, api_key=None, access_token=None, lang="en", version="v1", test=False):
+    def __init__(
+        self,
+        *,
+        api_key=None,
+        access_token=None,
+        lang="en",
+        version="v1",
+        test=False,
+        platform=None,
+        portal=None,
+    ):
         self.lang = lang
         self.version = version
         self.test = test
         self.connection = Connection(
-            test=test, api_key=api_key, access_token=access_token, version=version, lang=lang
+            test=test,
+            api_key=api_key,
+            access_token=access_token,
+            version=version,
+            lang=lang,
+            platform=platform,
+            portal=portal,
         )
 
     def __repr__(self):
@@ -287,6 +322,28 @@ class Client:
     @property
     def retry_after(self):
         return self.connection.retry_after
+
+    def set_platform(self, platform: Optional[TargetPlatform] = None) -> None:
+        """Change the platform targetted by the client. Call without an argument to not target any specific
+        paltform.
+
+        Parameters
+        -----------
+        platform : Optional[TargetPlatform]
+            The platform to set
+        """
+        self.connection.platform = platform
+
+    def set_portal(self, portal: Optional[TargetPortal] = None) -> None:
+        """Change the portal targetted by the client. Call without an argument to not target any specific
+        portal.
+
+        Parameters
+        -----------
+        portal : Optional[TargetPortal]
+            The portal to set
+        """
+        self.connection.portal = portal
 
     async def close(self):
         """|async| This function is used to clean up the client in order to close the application that it uses gracefully.